Ported UHHD demo
As part of the porting process I've converted this demo from a counter-based progression to wall timing both the effects themselves and the transitions.

Using time.time() means effects will run at a consistent speed no matter the speed of the platform. Running this demo on the Pi 4 on minicorn is hilariously fast otherwise.